+++ lzma/reader2.go
@@ -28,7 +28,7 @@ func (c *Reader2Config) fill() {
 // will be replaced by default values.
 func (c *Reader2Config) Verify() error {
 	c.fill()
-	if !(MinDictCap <= c.DictCap && c.DictCap <= MaxDictCap) {
+	if !(MinDictCap <= c.DictCap && int64(c.DictCap) <= MaxDictCap) {
 		return errors.New("lzma: dictionary capacity is out of range")
 	}
 	return nil
